# Conflicts:
#	src/main/java/com/keymanager/ckadmin/dao/CustomerKeywordDao.java
#	src/main/java/com/keymanager/ckadmin/service/CustomerKeywordService.java
#	src/main/resources/mybatis/ckmapper/CustomerKeywordMapper.xml
#	src/main/resources/mybatis/ckmapper/CustomerKeywordPositionSummaryMapper.xml

+++ src/main/java/com/keymanager/ckadmin/dao/NegativeListDao.java
@@ -17,7 +17,7 @@ public interface NegativeListDao extends BaseMapper<NegativeList> {
 
     List<NegativeList> getSpecifiedKeywordNegativeLists(@Param("keyword") String keyword);
 
-    int selectLastId();
+    long selectLastId();
 
     List<NegativeList> negativeListsSynchronizeOfDelete(@Param("negativeList") NegativeList negativeList);
